requirements.txt
mysql-connector-python
pymysql
cryptography
psutil
pywin32
wmi
requests
pyOpenSSL
paramiko
flask
flask-cors
python-dateutil
pytz
pyinstaller
certifi
charset-normalizer
idna
PyQt5
PyQt5-Qt5
PyQt5-sip
urllib3
